Move LastLocation and LocationCache to own module
